refactor(game/ui): move gamecard to entity layer
