+++ service/app/src/app/dashboard/components/dashboardGameSection.tsx
@@ -5,9 +5,9 @@ import { useIntersectionObserver } from "react-simplikit"
 
 import { useGamePreview } from "@/entities/game/hooks/useGamePreview"
 import { useInfiniteMyGames } from "@/entities/game/model/useInfiniteMyGames"
+import * as GameCard from "@/entities/game/ui/gameCard"
 import { GameCardOptions } from "@/entities/game/ui/gameCardOptions"
 import { GameLibrarySkeleton } from "@/entities/game/ui/gameLibrarySkeleton"
-import * as GameCard from "@/shared/gameCard"
 
 import { useDashboardGameActions } from "../hooks/useDashboardGameActions"
