+++ service/app/src/entities/game/ui/gameCard.tsx
